Create catalog-compile, start moving libraries around
