+++ lib/octocatalog-diff/api/bootstrap.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module OctocatalogDiff
+  module API
+    # This class allows octocatalog-diff to be used as a bootstrapping codeset.
+    class Bootstrap
+      #
+    end
+  end
+end
