+++ lib/octocatalog-diff/api/catalog-diff.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module OctocatalogDiff
+  module API
+    # This class allows octocatalog-diff to be used to obtain the difference between
+    # two catalogs.
+    class CatalogDiff
+      #
+    end
+  end
+end
