+++ spec/octocatalog-diff/integration/convert_file_resources_spec.rb
@@ -118,7 +118,7 @@ describe 'convert file resources' do
         ]
       )
       expect(result[:exitcode]).to eq(-1)
-      expect(result[:exception]).to be_a_kind_of(OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError)
+      expect(result[:exception]).to be_a_kind_of(OctocatalogDiff::Util::Catalogs::CatalogError)
       expect(result[:exception].message).to match(/failed to compile with Errno::ENOENT/)
       expect(result[:exception].message).to match(%r{Unable to resolve 'puppet:///modules/test/foo-new'})
     end
