+++ spec/octocatalog-diff/integration/preserve_environments_spec.rb
@@ -22,8 +22,8 @@ describe 'preserve environments integration' do
         expect(@result.exitcode).to eq(-1), OctocatalogDiff::Integration.format_exception(@result)
       end
 
-      it 'should raise OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError' do
-        expect(@result.exception).to be_a_kind_of(OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError)
+      it 'should raise OctocatalogDiff::Util::Catalogs::CatalogError' do
+        expect(@result.exception).to be_a_kind_of(OctocatalogDiff::Util::Catalogs::CatalogError)
       end
 
       it 'should fail because ::bar could not be located' do
@@ -98,7 +98,7 @@ describe 'preserve environments integration' do
 
       it 'should exit with error status due modules in production environment not being found' do
         expect(@result.exitcode).to eq(-1), OctocatalogDiff::Integration.format_exception(@result)
-        expect(@result.exception).to be_a_kind_of(OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError)
+        expect(@result.exception).to be_a_kind_of(OctocatalogDiff::Util::Catalogs::CatalogError)
         expect(@result.exception.message).to match(/Errno::ENOENT: No such file or directory - Environment directory/)
       end
     end
@@ -121,7 +121,7 @@ describe 'preserve environments integration' do
 
       it 'should error on missing environment' do
         expect(@result.exitcode).to eq(-1), OctocatalogDiff::Integration.format_exception(@result)
-        expect(@result.exception).to be_a_kind_of(OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError)
+        expect(@result.exception).to be_a_kind_of(OctocatalogDiff::Util::Catalogs::CatalogError)
         expect(@result.exception.message).to match(%r{Environment directory .+/environments/fluffy does not exist})
       end
     end
@@ -198,7 +198,7 @@ describe 'preserve environments integration' do
 
         it 'should error on missing site directory' do
           expect(@result.exitcode).to eq(-1), OctocatalogDiff::Integration.format_exception(@result)
-          expect(@result.exception).to be_a_kind_of(OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError)
+          expect(@result.exception).to be_a_kind_of(OctocatalogDiff::Util::Catalogs::CatalogError)
           expect(@result.exception.message).to match(/Could not find class (::)?sitetest/)
         end
       end
@@ -221,7 +221,7 @@ describe 'preserve environments integration' do
 
         it 'should error on missing module' do
           expect(@result.exitcode).to eq(-1), OctocatalogDiff::Integration.format_exception(@result)
-          expect(@result.exception).to be_a_kind_of(OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError)
+          expect(@result.exception).to be_a_kind_of(OctocatalogDiff::Util::Catalogs::CatalogError)
           expect(@result.exception.message).to match(/Could not find class (::)?foo/)
         end
       end
@@ -244,7 +244,7 @@ describe 'preserve environments integration' do
 
         it 'should raise exception due to missing symlink request' do
           expect(@result.exitcode).to eq(-1), OctocatalogDiff::Integration.format_exception(@result)
-          expect(@result.exception).to be_a_kind_of(OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError)
+          expect(@result.exception).to be_a_kind_of(OctocatalogDiff::Util::Catalogs::CatalogError)
           expect(@result.exception.message).to match(%r{Catalog for 'from' \(origin/master\) failed.+ Errno::ENOENT})
           expect(@result.exception.message).to match(%r{Specified directory .+/preserve-environments/fluffy doesn't exist})
         end
@@ -293,7 +293,7 @@ describe 'preserve environments integration' do
 
         it 'should error on missing site directory' do
           expect(@result.exitcode).to eq(-1), OctocatalogDiff::Integration.format_exception(@result)
-          expect(@result.exception).to be_a_kind_of(OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError)
+          expect(@result.exception).to be_a_kind_of(OctocatalogDiff::Util::Catalogs::CatalogError)
           expect(@result.exception.message).to match(/Could not find class (::)?sitetest/)
         end
       end
@@ -316,7 +316,7 @@ describe 'preserve environments integration' do
 
       it 'should error on missing site directory' do
         expect(@result.exitcode).to eq(-1), OctocatalogDiff::Integration.format_exception(@result)
-        expect(@result.exception).to be_a_kind_of(OctocatalogDiff::CatalogDiff::Cli::Catalogs::CatalogError)
+        expect(@result.exception).to be_a_kind_of(OctocatalogDiff::Util::Catalogs::CatalogError)
         expect(@result.exception.message).to match(/Could not find class (::)?sitetest/)
       end
     end
